Revert very recent backport of changes to the type system
The backport of the change made for PR c/113688 onto the 14 branch a couple
of weeks ago has seriously broken the LTO compiler for the Ada language on
the 14 branch, because it changes the GCC type system for the sake of C in
a way that is not compatible with simple discriminated types in Ada.  To be
more precise, useless_type_conversion_p now returns true for some (view-)
conversions that are needed by the rest of the compiler.

gcc/
	PR lto/119792
	Revert

	Backported from master:
	    2024-12-12  Martin Uecker  <uecker@tugraz.at>

	PR c/113688
	PR c/114014
	PR c/114713
	PR c/117724
	* tree.cc (gimple_canonical_types_compatible_p): Add exception.
	(verify_type): Add exception.

gcc/lto/
	PR lto/119792
	Revert

	Backported from master:
	    2024-12-12  Martin Uecker  <uecker@tugraz.at>
	* lto-common.cc (hash_canonical_type): Add exception.

gcc/testsuite/
	* gcc.dg/pr113688.c: Delete.
	* gcc.dg/pr114014.c: Likewise.
	* gcc.dg/pr114713.c: Likewise.
	* gcc.dg/pr117724.c: Likewise