Set znver5 issue rate to 4.
this patch sets issue rate of znver5 to 4.  With current model, unless a reservation is
missing, we will never issue more than 4 instructions per cycle since that is the limit
of decoders and the model does not take into acount the fact that typically code is run
from op cache.

gcc/ChangeLog:

	* config/i386/x86-tune-sched.cc (ix86_issue_rate): Set
	to 4 for znver5.